+++ packages/elements/src/icon/index.ts
@@ -14,6 +14,8 @@ import { Deferred } from '@refinitiv-ui/utils/loader.js';
 import { VERSION } from '../version.js';
 import { IconLoader } from './utils/IconLoader.js';
 export { preload } from './utils/IconLoader.js';
+import { consume } from '@lit-labs/context';
+import { efConfig, type Config } from '../configuration/index.js';
 
 const EmptyTemplate = svg``;
 
@@ -34,6 +36,13 @@ export class Icon extends BasicElement {
   static get version (): string {
     return VERSION;
   }
+
+  /**
+   * Icon map from ef-configuration
+   * @ignore
+   */
+  @consume({ context: efConfig, subscribe: true })
+  public config?: Config;
 
   /**
    * A `CSSResultGroup` that will be used
@@ -60,7 +69,7 @@ export class Icon extends BasicElement {
 
   /**
    * Name of a known icon to render or URL of SVG icon.
-   * @example home | https://cdn.io/icons/home.svg
+   * @example heart | https://cdn.io/icons/heart.svg
    * @default null
    */
   @property({ type: String, reflect: true })
@@ -82,7 +91,7 @@ export class Icon extends BasicElement {
   /**
    * Src location of an svg icon.
    * @ignore
-   * @example https://cdn.io/icons/home.svg
+   * @example https://cdn.io/icons/heart.svg
    * @deprecated Use `icon` instead
    * @default null
    */
@@ -91,14 +100,17 @@ export class Icon extends BasicElement {
     return this._src;
   }
   /**
-    * @ignore
-    * @param value Location of an svg
-    */
+   * @ignore
+   * @param value Location of an svg
+   */
   public set src (value: string | null) {
     if (this.src !== value) {
       this.deferIconReady();
       this._src = value;
-      if (value) {
+      if (this.icon && this.iconMap) {
+        void this.loadAndRenderIcon(this.iconMap);
+      }
+      else if (value) {
         void this.loadAndRenderIcon(value);
       }
       else {
@@ -136,6 +148,14 @@ export class Icon extends BasicElement {
     // `iconReady` resolves at this stage so that `updateComplete` would be resolvable
     // even in the case that both `icon` and `src` attribute are missing.
     this.iconReady.resolve();
+  }
+
+  /**
+   * Check if the icon map configuration has content
+   * @returns icon map if exists
+   */
+  private get iconMap (): string | null {
+    return this.icon && this.config?.icon.map[this.icon] || null;
   }
 
   /**